atualizando projeto 05/11/23

+++ springboot-rest-api-sample/src/main/java/br/com/jumbo/projeto_insulina/controllers/UsuarioController.java
@@ -42,7 +42,7 @@ public class UsuarioController {
 
 	@Autowired
 	UsuarioService usuarioService;
-	
+
 	@ResponseBody
 	@PostMapping(value = "**/salvarUsuario")
 	public ResponseEntity<Usuario> salvarUsuario(@RequestBody @Valid Usuario usuario) throws ExceptionProjetoInsulina {
@@ -60,46 +60,21 @@ public class UsuarioController {
 				throw new ExceptionProjetoInsulina(
 						"O email: " + usuario.getEmail() + " já está cadastrado no Banco de Dados");
 			}
-			
-		}if(usuario.getTipoSenha().getDescricao().equalsIgnoreCase("SIMPLES")){
-			
+
+		}
+		if (usuario.getTipoSenha().getDescricao().equalsIgnoreCase("SIMPLES")) {
+
 			Usuario usuario1 = usuarioService.salvaUsuario(usuario);
-		//	return new ResponseEntity<Usuario>(usuario1, HttpStatus.OK);
-			
-		}if(usuario.getTipoSenha().getDescricao().equalsIgnoreCase("CRYPTO")) {
-			
+
+		}
+		if (usuario.getTipoSenha().getDescricao().equalsIgnoreCase("CRYPTO")) {
+
 			Usuario usuario1 = usuarioService.salvaSenhaCriptUsuario(usuario);
-		//	return new ResponseEntity<Usuario>(usuario1, HttpStatus.OK);
+
 		}
-           
-		
+
 		return new ResponseEntity<Usuario>(usuario, HttpStatus.OK);
-		
-	}
-	
 
-	@ResponseBody
-	@PostMapping(value = "**/salvarUsuarioSenhaCript")
-	public ResponseEntity<Usuario> salvarUsuarioSenhaCript(@RequestBody @Valid Usuario usuario) throws ExceptionProjetoInsulina {
-
-		if (usuario.getId() == null) {
-			List<Usuario> usuarios = usuarioRepository.buscaUsuarioLogin(usuario.getLogin().toUpperCase());
-
-			if (!usuarios.isEmpty()) {
-
-				throw new ExceptionProjetoInsulina("Já exixte Login com essa descrição: " + usuario.getLogin());
-
-			}
-			List<Usuario> usuarios1 = usuarioRepository.buscaUsuarioEmail(usuario.getEmail().toUpperCase());
-			if (!usuarios1.isEmpty()) {
-				throw new ExceptionProjetoInsulina(
-						"O email: " + usuario.getEmail() + " já está cadastrado no Banco de Dados");
-			}
-		}
-
-		Usuario usuario1 = usuarioService.salvaSenhaCriptUsuario(usuario);
-
-		return new ResponseEntity<Usuario>(usuario1, HttpStatus.OK);
 	}
 
 	@ResponseBody
@@ -116,6 +91,13 @@ public class UsuarioController {
 	@DeleteMapping(value = "**/deleteUsuarioPorId/{id}")
 	public ResponseEntity<?> deleteUsuarioPorId(@PathVariable("id") Long id) throws ExceptionProjetoInsulina {
 
+		List<Usuario> usuarios1 = usuarioRepository.buscaUsuarioId(id);
+		if (usuarios1.isEmpty()) {
+			throw new ExceptionProjetoInsulina(
+					"O Usuário com Id: " + id + " não existe no Banco de Dados");
+		}
+
+	
 		usuarioRepository.deleteById(id);
 
 		return new ResponseEntity("Usuario deletado por Id com sucesso!", HttpStatus.OK);
@@ -136,25 +118,26 @@ public class UsuarioController {
 		return new ResponseEntity<Usuario>(usuario, HttpStatus.OK);
 
 	}
-	
+
 	@ResponseBody
 	@GetMapping(value = "**/buscaUsuarioPorLogin/{login}")
-	public ResponseEntity<List<Usuario>> buscaUsuarioPorDesc(@PathVariable("login") String login)throws ExceptionProjetoInsulina {
+	public ResponseEntity<List<Usuario>> buscaUsuarioPorDesc(@PathVariable("login") String login)
+			throws ExceptionProjetoInsulina {
 
 		List<Usuario> usuario = usuarioRepository.buscaUsuarioPorDesc(login.toUpperCase());
-	
+
 		if (usuario.isEmpty()) {
 
 			throw new ExceptionProjetoInsulina("O Usuario: " + login + ", não contém no banco de dados. ");
 		}
 
-
 		return new ResponseEntity<List<Usuario>>(usuario, HttpStatus.OK);
 	}
-	
+
 	@ResponseBody
 	@GetMapping(value = "**/buscaUsuarioPorNome/{nome}")
-	public ResponseEntity<List<Usuario>> buscaUsuarioPorNome(@PathVariable("nome") String nome) throws ExceptionProjetoInsulina {
+	public ResponseEntity<List<Usuario>> buscaUsuarioPorNome(@PathVariable("nome") String nome)
+			throws ExceptionProjetoInsulina {
 
 		List<Usuario> usuarios = usuarioRepository.buscaUsuarioPorNome(nome.toUpperCase());
 
@@ -162,92 +145,85 @@ public class UsuarioController {
 
 			throw new ExceptionProjetoInsulina("Não encotrado Usuário com o nome: " + nome);
 		}
-		
+
 		return new ResponseEntity<List<Usuario>>(usuarios, HttpStatus.OK);
 	}
-	
+
 	@ResponseBody
 	@GetMapping(value = "**/buscaUsuarioPorEmail/{email}")
-	public ResponseEntity<List<Usuario>> buscaUsuarioPorEmail(@PathVariable("email") String nome) throws ExceptionProjetoInsulina {
+	public ResponseEntity<List<Usuario>> buscaUsuarioPorEmail(@PathVariable("email") String nome)
+			throws ExceptionProjetoInsulina {
 
 		List<Usuario> usuarios = usuarioRepository.buscaUsuarioPorEmail(nome.toUpperCase());
 
 		if (usuarios.isEmpty()) {
 
-			throw new ExceptionProjetoInsulina("Email não cadastrado no BD" );
+			throw new ExceptionProjetoInsulina("Email não cadastrado no BD");
 		}
-		
+
 		return new ResponseEntity<List<Usuario>>(usuarios, HttpStatus.OK);
 	}
-	
+
 //@ResponseBody
 //	@GetMapping(value = "**/buscaUsuarioPorLogin/{login}")
- //  public ResponseEntity<List<Usuario>> buscaUsuarioPorDesc(@PathVariable("login") String login)throws ExceptionProjetoInsulina {
+	// public ResponseEntity<List<Usuario>>
+	// buscaUsuarioPorDesc(@PathVariable("login") String login)throws
+	// ExceptionProjetoInsulina {
 
-		//List<Usuario> usuario = usuarioRepository.buscaUsuarioPorDesc(login.toUpperCase());
-	
-	//	if (usuario.isEmpty()) {
+	// List<Usuario> usuario =
+	// usuarioRepository.buscaUsuarioPorDesc(login.toUpperCase());
 
-		//	throw new ExceptionProjetoInsulina("O Usuario: " + login + ", não contém no banco de dados. ");
-	//	}
+	// if (usuario.isEmpty()) {
 
+	// throw new ExceptionProjetoInsulina("O Usuario: " + login + ", não contém no
+	// banco de dados. ");
+	// }
 
-		//return new ResponseEntity<List<Usuario>>(usuario, HttpStatus.OK);
-	//}
-	
+	// return new ResponseEntity<List<Usuario>>(usuario, HttpStatus.OK);
+	// }
+
 	@ResponseBody
 	@GetMapping(value = "**/logarUsuario/{login}")
-	 public ResponseEntity<List<Usuario>> logarUsuario(@PathVariable("login") String login)throws ExceptionProjetoInsulina {
-		
+	public ResponseEntity<List<Usuario>> logarUsuario(@PathVariable("login") String login)
+			throws ExceptionProjetoInsulina {
+
 		List<Usuario> usuario = usuarioRepository.buscaUsuarioPorDesc(login.toUpperCase());
-		
-			if (usuario.isEmpty()) {
-				
-				
 
-				throw new ExceptionProjetoInsulina("O Usuario: " + login + ", não contém no banco de dados. ");
-			
-			
-			}
+		if (usuario.isEmpty()) {
 
+			throw new ExceptionProjetoInsulina("O Usuario: " + login + ", não contém no banco de dados. ");
 
-			return new ResponseEntity<List<Usuario>>(usuario, HttpStatus.OK);
 		}
-			
-		//	throw new ExceptionProjetoInsulina("Não encotrado Usuario com código " + login);
-			
-		
-	//	}else if(usuario != null) {
-			
-		//	Long numero = (long) 37;
-		//	String senha1 = "123";
-			   
-	
-			
-			
-	
-				
-			//	throw new ExceptionProjetoInsulina("Senha confere com o número digitado ");
-				
-			//}else {
-			//	throw new ExceptionProjetoInsulina("Senha não confere com o número digitado ");
-		//	}
-			
-	//throw new ExceptionProjetoInsulina("Usuario encotrado " + usuario.getNome() + " login: " + usuario.getLogin()
-			//+ " senha: "+usuario.getSenha() + " email: " + usuario.getEmail());
-			
-			
-			
-			 //usuario = UsuarioService.consultarUsuario(id);
-			 
-			
-		
-			//throw new ExceptionProjetoInsulina("usuario encontrado:  " + usuario.getId());
-	
 
-		//return new ResponseEntity<Usuario>(usuario, HttpStatus.OK);
+		return new ResponseEntity<List<Usuario>>(usuario, HttpStatus.OK);
+	}
 
-	//}
-		//return new ResponseEntity<Usuario>(usuario, HttpStatus.OK);
-       //   }
+	// throw new ExceptionProjetoInsulina("Não encotrado Usuario com código " +
+	// login);
+
+	// }else if(usuario != null) {
+
+	// Long numero = (long) 37;
+	// String senha1 = "123";
+
+	// throw new ExceptionProjetoInsulina("Senha confere com o número digitado ");
+
+	// }else {
+	// throw new ExceptionProjetoInsulina("Senha não confere com o número digitado
+	// ");
+	// }
+
+	// throw new ExceptionProjetoInsulina("Usuario encotrado " + usuario.getNome() +
+	// " login: " + usuario.getLogin()
+	// + " senha: "+usuario.getSenha() + " email: " + usuario.getEmail());
+
+	// usuario = UsuarioService.consultarUsuario(id);
+
+	// throw new ExceptionProjetoInsulina("usuario encontrado: " + usuario.getId());
+
+	// return new ResponseEntity<Usuario>(usuario, HttpStatus.OK);
+
+	// }
+	// return new ResponseEntity<Usuario>(usuario, HttpStatus.OK);
+	// }
 }
